Fix Nft Integration Tests Use only one BridgeFactory for both Bridge and NftBridge as they share the same factory from Federator.ts runAll Update Cross event in the NftBridge to filter per destinationChainId
