+++ bridge/abi/NFTBridge.json
@@ -152,26 +152,32 @@
       {
         "indexed": true,
         "internalType": "address",
-        "name": "_from",
+        "name": "_to",
         "type": "address"
       },
       {
         "indexed": true,
+        "internalType": "uint256",
+        "name": "_destinationChainId",
+        "type": "uint256"
+      },
+      {
+        "indexed": false,
         "internalType": "address",
-        "name": "_to",
+        "name": "_from",
         "type": "address"
+      },
+      {
+        "indexed": false,
+        "internalType": "uint256",
+        "name": "_originChainId",
+        "type": "uint256"
       },
       {
         "indexed": false,
         "internalType": "address",
         "name": "_tokenCreator",
         "type": "address"
-      },
-      {
-        "indexed": false,
-        "internalType": "bytes",
-        "name": "_userData",
-        "type": "bytes"
       },
       {
         "indexed": false,
@@ -193,15 +199,9 @@
       },
       {
         "indexed": false,
-        "internalType": "uint256",
-        "name": "_originChainId",
-        "type": "uint256"
-      },
-      {
-        "indexed": false,
-        "internalType": "uint256",
-        "name": "_destinationChainId",
-        "type": "uint256"
+        "internalType": "bytes",
+        "name": "_userData",
+        "type": "bytes"
       }
     ],
     "name": "Cross",
